+++ src/main/java/tech/jazz/apianalisecredito/infrastructure/model/CreditAnalysis.java
@@ -1,22 +1,22 @@
 package tech.jazz.apianalisecredito.infrastructure.model;
 
-import lombok.Builder;
-
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
+import lombok.Builder;
 
 public record CreditAnalysis(
         Boolean approved,
         BigDecimal approvedLimit,
         BigDecimal requestedAmount,
         BigDecimal withdraw,
-        Integer annualInterest,
+        Float annualInterest,
         UUID clientId,
         LocalDateTime date
 ) {
     @Builder
-    public CreditAnalysis(Boolean approved, BigDecimal approvedLimit, BigDecimal requestedAmount, BigDecimal withdraw, Integer annualInterest, UUID clientId, LocalDateTime date) {
+    public CreditAnalysis(Boolean approved, BigDecimal approvedLimit, BigDecimal
+            requestedAmount, BigDecimal withdraw, Float annualInterest, UUID clientId, LocalDateTime date) {
         this.approved = approved;
         this.approvedLimit = approvedLimit;
         this.requestedAmount = requestedAmount;
